Unserializable Return Type
When an unserializable type is returned such as a function, this results in an uncaught error. In Deno, this crashes the application. This change catches the error and returns a serialized error type so that it can be handled by the calling thread.

Created unit test for handling unserializable types