Improve two-way communication test
This commit improves the test to capture clone errors that occured in v4.3.1
(or lower) when trying to do two-way communication.

To do this, we expose a NonCloneable object that contains a MessagePort.
In v4.3.1 when `proxy` is called, the listener attached by `Comlink.expose`
also gets called by the remote and does not perform an early return.

Instead it attempt to do a postMessage with the raw value set to the
NonCloneable instance.

This throw an unhandledrejection.

With this new test we verify that this does not happen.