Support multiple format to allow communication with 4.3.0 or older
The approach taken is a bit different from the previous commit.
We now support type values to be either the format from 4.3.0 or
the format introduced in 4.3.1.

When communicating with 4.3.0, the only way to communicate is if
it's done like this:

   // 4.3.0                     ------------>         latest

   const a = Comlink.wrap(ep)                   Comlink.expose(b, ep)

This works because `a` has to initiate the conversation with `b`
giving us a chance to always treat future messages as 4.3.0 ones.

In particular we support proxy sent by `a`:

   a.registerCallback(Comlink.proxy(() => {}))

But we also support sending MessagePort that can be later wrapped
by the other side:

   a.callMeBack(Comlink.transfer(port, [port]))